backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
pillow==10.2.0
mysqlclient==2.1.1
django-filter==23.5